+++ coop/tsmod/zb_flood.c
@@ -74,7 +74,7 @@ qboolean ReadFloodFile(char *floodname)
 			uptoLine++;
 			
 			// remove '\n'
-			len = q2a_strlen(buffer) - 1;
+			len = (int)q2a_strlen(buffer) - 1;
 			if(buffer[len] == '\n')
 				{
 					buffer[len] = 0x0;
@@ -103,7 +103,7 @@ qboolean ReadFloodFile(char *floodname)
 					cp += 3;
 					SKIPBLANK(cp);
 					
-					len = q2a_strlen(cp) + 1;
+					len = (int)q2a_strlen(cp) + 1;
 					
 					// zero length command
 					if(!len)
@@ -144,7 +144,7 @@ qboolean ReadFloodFile(char *floodname)
 				}
 			else if(!(cp[0] == ';' || cp[0] == '\n' || isBlank (cp)))
 				{
-					gi.dprintf("Error loading FLOOD from line %d in file %s\n", uptoLine, floodname);
+					gi.dprintf ("Error loading FLOOD from line %d in file %s\n", uptoLine, floodname);
 				}
 		}
 		
@@ -185,7 +185,7 @@ void readFloodLists(void)
 		
 	if(!ret)
 		{
-			gi.dprintf("WARNING: " FLOODFILE " could not be found\n");
+			gi.dprintf ("WARNING: " FLOODFILE " could not be found\n");
 			logEvent(LT_INTERNALWARN, 0, NULL, FLOODFILE " could not be found", IW_FLOODSETUPLOAD, 0.0);
 		}
 }
@@ -790,7 +790,7 @@ void floodcmdRun(int startarg, edict_t *ent, int client)
 			return;
 		}
 		
-	len = q2a_strlen(cmd) + 20;
+	len = (int)q2a_strlen(cmd) + 20;
 	
 	floodcmds[maxflood_cmds].floodcmd = gi.TagMalloc (len, TAG_LEVEL);
 	processstring(floodcmds[maxflood_cmds].floodcmd, cmd, len - 1, 0);
@@ -871,7 +871,7 @@ void floodDelRun(int startarg, edict_t *ent, int client)
 		
 	if(flood + 1 < maxflood_cmds)
 		{
-			q2a_memmove((floodcmds + flood), (floodcmds + flood + 1), sizeof(floodcmd_t) * (maxflood_cmds - flood));
+			q2a_memmove((floodcmds + flood), (floodcmds + flood + 1), sizeof(floodcmd_t) * ((size_t)maxflood_cmds - flood));
 		}
 		
 	maxflood_cmds--;
